refactor(generator): add dedicated stage with workflow for data fiddling
the data from the nautobot needs to be fiddled a bit before it can
be used in the render. for example, a serial number and hostname gets
defined for every device. additionally, a usecase string is assembled
from the device role, type, and manufacturer. depending on the usecase
a different template gets selected, which may requires changes to the
context.

co-authored by raven <raven@mwsq.de>